rename and update logic in emailAuth
